pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "doc2json"
version = "0.1.0"
description = "Turn unstructured documents into clean JSON with auto-generated schemas"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = [
    "document-extraction",
    "pdf",
    "json",
    "llm",
    "pydantic",
    "structured-data",
    "ocr",
    "data-extraction",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "instructor>=1.0",
    "python-dotenv>=1.0",
    # Parsers - all included in core
    "pdfplumber>=0.10",
    "pdf2image>=1.16",
    "pytesseract>=0.3",
    "python-docx>=1.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.scripts]
doc2json = "doc2json.cli.main:cli"

[project.optional-dependencies]
# LLM providers - install the one(s) you need
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.5"]
all = [
    "anthropic>=0.40",
    "openai>=1.0",
    "google-generativeai>=0.5",
]
# Connectors - optional source/destination integrations
google-drive = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
]
postgres = [
    "psycopg2-binary>=2.9",
]
mongodb = [
    "pymongo>=4.0",
]
sql = [
    "sqlalchemy>=2.0",
]
snowflake = [
    "snowflake-connector-python>=3.0",
]
bigquery = [
    "google-cloud-bigquery>=3.0",
]
s3 = [
    "boto3>=1.34",
]
azure-blob = [
    "azure-storage-blob>=12.0",
]
dev = [
    "pytest>=7.0",
]

[project.urls]
Homepage = "https://github.com/DataFenix-Ltd/doc2json"
Repository = "https://github.com/DataFenix-Ltd/doc2json"
Issues = "https://github.com/DataFenix-Ltd/doc2json/issues"

[tool.setuptools.packages.find]
include = ["doc2json*"]
